changes UI if iframe
