+++ sandbox-iframe/src/components/NoParent.tsx
@@ -1,0 +1,5 @@
+export function NoParent() {
+  return (
+    <div></div>
+  )
+}
